+++ tests/functional/lt_test_rev_ping.c
@@ -6,6 +6,7 @@
  * @license For the license see file LICENSE.txt file in the root directory of this source tree.
  */
 
+#include "inttypes.h"
 #include "libtropic.h"
 #include "libtropic_common.h"
 #include "libtropic_functional_tests.h"
@@ -34,22 +35,22 @@ void lt_test_rev_ping(void)
     LT_LOG_INFO("Initializing handle");
     LT_TEST_ASSERT(LT_OK, lt_init(&h));
 
-    LT_LOG_INFO("Starting Secure Session with key %d", PAIRING_KEY_SLOT_INDEX_0);
+    LT_LOG_INFO("Starting Secure Session with key %d", (int)PAIRING_KEY_SLOT_INDEX_0);
     LT_TEST_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh0priv, sh0pub, PAIRING_KEY_SLOT_INDEX_0));
     LT_LOG_LINE();
 
     LT_LOG_INFO("Will send %d Ping commands with random data of random length", PING_MAX_LOOPS);
     for (uint16_t i = 0; i < PING_MAX_LOOPS; i++) {
         LT_LOG_INFO();
-        LT_LOG_INFO("Generating random data length <= %d...", PING_LEN_MAX);
+        LT_LOG_INFO("Generating random data length <= %d...", (int)PING_LEN_MAX);
         LT_TEST_ASSERT(LT_OK, lt_port_random_bytes(&random_data_size, 1));
         random_data_size %= PING_LEN_MAX + 1;  // 0-4096
 
-        LT_LOG_INFO("Generating %d random bytes...", random_data_size);
+        LT_LOG_INFO("Generating %" PRIu32 " random bytes...", random_data_size);
         LT_TEST_ASSERT(LT_OK, lt_port_random_bytes(random_data, sizeof(random_data) / sizeof(uint32_t)));
         memcpy(ping_msg_out, random_data, random_data_size);
 
-        LT_LOG_INFO("Sending Ping command #%d...", i);
+        LT_LOG_INFO("Sending Ping command #%" PRIu16 "...", i);
         LT_TEST_ASSERT(LT_OK, lt_ping(&h, ping_msg_out, ping_msg_in, random_data_size));
 
         LT_LOG_INFO("Comparing sent and received message...");
