+++ tests/functional/lt_test_rev_r_mem.c
@@ -6,6 +6,7 @@
  * @license For the license see file LICENSE.txt file in the root directory of this source tree.
  */
 
+#include "inttypes.h"
 #include "libtropic.h"
 #include "libtropic_common.h"
 #include "libtropic_functional_tests.h"
@@ -22,7 +23,7 @@ lt_ret_t lt_test_rev_r_mem_cleanup(void)
     uint8_t r_mem_data[R_MEM_DATA_SIZE_MAX];
     uint16_t read_data_size;
 
-    LT_LOG_INFO("Starting secure session with slot %d", PAIRING_KEY_SLOT_INDEX_0);
+    LT_LOG_INFO("Starting secure session with slot %d", (int)PAIRING_KEY_SLOT_INDEX_0);
     ret = verify_chip_and_start_secure_session(&h, sh0priv, sh0pub, PAIRING_KEY_SLOT_INDEX_0);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to establish secure session.");
@@ -32,14 +33,14 @@ lt_ret_t lt_test_rev_r_mem_cleanup(void)
     LT_LOG_INFO("Erasing all slots...");
     for (uint16_t i = 0; i <= R_MEM_DATA_SLOT_MAX; i++) {
         LT_LOG_INFO();
-        LT_LOG_INFO("Erasing slot #%d...", i);
+        LT_LOG_INFO("Erasing slot #%" PRIu16 "...", i);
         ret = lt_r_mem_data_erase(&h, i);
         if (LT_OK != ret) {
             LT_LOG_ERROR("Failed to erase slot.");
             return ret;
         }
 
-        LT_LOG_INFO("Reading slot #%d (should fail)...", i);
+        LT_LOG_INFO("Reading slot #%" PRIu16 " (should fail)...", i);
         ret = lt_r_mem_data_read(&h, i, r_mem_data, &read_data_size);
         if (LT_L3_R_MEM_DATA_READ_SLOT_EMPTY != ret) {
             LT_LOG_ERROR("Return value is not LT_L3_R_MEM_DATA_READ_SLOT_EMPTY.");
@@ -88,14 +89,14 @@ void lt_test_rev_r_mem(void)
     LT_LOG_INFO("Initializing handle");
     LT_TEST_ASSERT(LT_OK, lt_init(&h));
 
-    LT_LOG_INFO("Starting Secure Session with key %d", PAIRING_KEY_SLOT_INDEX_0);
+    LT_LOG_INFO("Starting Secure Session with key %d", (int)PAIRING_KEY_SLOT_INDEX_0);
     LT_TEST_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh0priv, sh0pub, PAIRING_KEY_SLOT_INDEX_0));
     LT_LOG_LINE();
 
     LT_LOG_INFO("Checking if all slots are empty...");
     for (uint16_t i = 0; i <= R_MEM_DATA_SLOT_MAX; i++) {
         LT_LOG_INFO();
-        LT_LOG_INFO("Reading slot #%d (should fail)...", i);
+        LT_LOG_INFO("Reading slot #%" PRIu16 " (should fail)...", i);
         LT_TEST_ASSERT(LT_L3_R_MEM_DATA_READ_SLOT_EMPTY, lt_r_mem_data_read(&h, i, r_mem_data, &read_data_size));
 
         LT_LOG_INFO("Checking number of read bytes (should be 0)...");
@@ -109,14 +110,14 @@ void lt_test_rev_r_mem(void)
     LT_LOG_INFO("Testing writing all slots entirely...");
     for (uint16_t i = 0; i <= R_MEM_DATA_SLOT_MAX; i++) {
         LT_LOG_INFO();
-        LT_LOG_INFO("Generating random data for slot #%d...", i);
+        LT_LOG_INFO("Generating random data for slot #%" PRIu16 "...", i);
         LT_TEST_ASSERT(LT_OK, lt_port_random_bytes(random_data, sizeof(random_data) / sizeof(uint32_t)));
         memcpy(write_data, random_data, sizeof(write_data));
 
-        LT_LOG_INFO("Writing to slot #%d...", i);
+        LT_LOG_INFO("Writing to slot #%" PRIu16 "...", i);
         LT_TEST_ASSERT(LT_OK, lt_r_mem_data_write(&h, i, write_data, R_MEM_DATA_SIZE_MAX));
 
-        LT_LOG_INFO("Reading slot #%d...", i);
+        LT_LOG_INFO("Reading slot #%" PRIu16 "...", i);
         LT_TEST_ASSERT(LT_OK, lt_r_mem_data_read(&h, i, r_mem_data, &read_data_size));
 
         LT_LOG_INFO("Checking number of read bytes...");
@@ -125,10 +126,10 @@ void lt_test_rev_r_mem(void)
         LT_LOG_INFO("Checking contents...");
         LT_TEST_ASSERT(0, memcmp(r_mem_data, write_data, R_MEM_DATA_SIZE_MAX));
 
-        LT_LOG_INFO("Writing zeros to slot #%d (should fail)...", i);
+        LT_LOG_INFO("Writing zeros to slot #%" PRIu16 " (should fail)...", i);
         LT_TEST_ASSERT(LT_L3_R_MEM_DATA_WRITE_WRITE_FAIL, lt_r_mem_data_write(&h, i, zeros, R_MEM_DATA_SIZE_MAX));
 
-        LT_LOG_INFO("Reading slot #%d...", i);
+        LT_LOG_INFO("Reading slot #%" PRIu16 "...", i);
         read_data_size = 0;  // Set different value just in case
         LT_TEST_ASSERT(LT_OK, lt_r_mem_data_read(&h, i, r_mem_data, &read_data_size));
 
@@ -143,10 +144,10 @@ void lt_test_rev_r_mem(void)
     LT_LOG_INFO("Erasing all slots...");
     for (uint16_t i = 0; i <= R_MEM_DATA_SLOT_MAX; i++) {
         LT_LOG_INFO();
-        LT_LOG_INFO("Erasing slot #%d...", i);
+        LT_LOG_INFO("Erasing slot #%" PRIu16 "...", i);
         LT_TEST_ASSERT(LT_OK, lt_r_mem_data_erase(&h, i));
 
-        LT_LOG_INFO("Reading slot #%d (should fail)...", i);
+        LT_LOG_INFO("Reading slot #%" PRIu16 " (should fail)...", i);
         LT_TEST_ASSERT(LT_L3_R_MEM_DATA_READ_SLOT_EMPTY, lt_r_mem_data_read(&h, i, r_mem_data, &read_data_size));
 
         LT_LOG_INFO("Checking number of read bytes (should be 0)...");
@@ -157,19 +158,19 @@ void lt_test_rev_r_mem(void)
     LT_LOG_INFO("Testing writing all slots partially or with zero data size...");
     for (uint16_t i = 0; i <= R_MEM_DATA_SLOT_MAX; i++) {
         LT_LOG_INFO();
-        LT_LOG_INFO("Generating random data length < %d...", R_MEM_DATA_SIZE_MAX);
+        LT_LOG_INFO("Generating random data length < %d...", (int)R_MEM_DATA_SIZE_MAX);
         LT_TEST_ASSERT(LT_OK, lt_port_random_bytes(&random_data_size, 1));
         random_data_size %= R_MEM_DATA_SIZE_MAX;
 
-        LT_LOG_INFO("Generating %d random bytes for slot #%d...", random_data_size, i);
+        LT_LOG_INFO("Generating %" PRIu32 " random bytes for slot #%" PRIu16 "...", random_data_size, i);
         LT_TEST_ASSERT(LT_OK, lt_port_random_bytes(random_data, sizeof(random_data) / sizeof(uint32_t)));
         memcpy(write_data, random_data, random_data_size);
 
-        LT_LOG_INFO("Writing to slot #%d...", i);
+        LT_LOG_INFO("Writing to slot #%" PRIu16 "...", i);
         LT_TEST_ASSERT_COND(lt_r_mem_data_write(&h, i, write_data, random_data_size), random_data_size != 0, LT_OK,
                             LT_L3_FAIL);
 
-        LT_LOG_INFO("Reading slot #%d...", i);
+        LT_LOG_INFO("Reading slot #%" PRIu16 "...", i);
         LT_TEST_ASSERT_COND(lt_r_mem_data_read(&h, i, r_mem_data, &read_data_size), random_data_size != 0, LT_OK,
                             LT_L3_R_MEM_DATA_READ_SLOT_EMPTY);
